Add relations code, refactor static stuff, change sql code to module
This will add a page to get database relations as a json object.

This also changes the sql stuff from being a separate node app to a
module that gets required in the main app (also changed the urls a  bit)

And finally, this moves static stuff out of the views folder and in to
a static folder. Since we have no views anymore, this also means the
views folder is empty. The html and angular code was also changed so
everything isn't prefixed with 'views/'


Former-commit-id: 0c41987e2e7a5b74e13298543c73b472d6af182e
Former-commit-id: 8b3744bac2d9db393bcd13c09ff3fa0b758fa4b7