Update mysql code for GoJS and to use database connection info
This changes the mysql code to use POST with the database connection
info as the body of the request. It also changes the get data function
to return database schema.

The JSON returned by both SQL functions is also now much more friendly
for GoJS, so the processing we have to do on it is very minimal.


Former-commit-id: f32d73df394da06774ed1fe41f8e2f98fc4a650b
Former-commit-id: 97f4da1453a0a0234a5326482c44374b667b4400