DROP-21, DROP-20, DROP-8, DROP-9: Link up database with project view
This finishes up most of our code. It fixes some bugs in the mysql
queries, and links up the GoJS code with the mysql JSON objects, which
are fetched by HTTP calls through Angular.

This also adds a Database parameter to projects, as we need to know
which database on the database server to look at ('employees' for the
test database).


Former-commit-id: 3745f65dc4ddf26733116e96cac98bd836262e4d
Former-commit-id: b840fd34659770da3e85c2950be107c64fe698db