Add abstract relations code, make sql code more ES6-ey
This adds the abstract relation creation code, based on the pseudo-code
provided to us. I had to make some fairly significant changes since a
the pseudo-code seems to have a lot of bugs, but it seems to be working
now. Note that it doesn't actually do anything right now, but if you
debug it and inspect the abstractEntities, abstractRelations, and
argument variables at the end of the group function, they seem to have
the proper values.

Also changed some of the sql code to use ES6 since Node has full support
for ES6 and it makes some stuff cleaner.