Add initial work for JavaParser
This adds the basic code we will need for JavaParser. Currently it is
able to find entities, find relations, and figure out full class names
of relation targets. That is pretty much all we need as far as parsing
goes, so all that is left is to format it all nicely and return it as
a JSON object for Node to read.